clean code
